#768: Update environment (#774)
#768 Environment updates including;
 * python 3.12 plus updated shebangs.
 * CMOR v3.13.2
 * Tweaks to cdds unit tests.
 * Move mkdocs from pip section to conda section.
 * Add a couple xfails on tests for future investigation.
 * Add some tolerance values to prevent tests failing after numpy & python update.
 * Add COMBINE_POLICY pp loading fix to improve iris performance
 * Add flag to mip_convert to silence iris microseconds warnings.

+++ mip_convert/mip_convert/plugins/base/data/processors.py
@@ -949,12 +949,12 @@ def eos_insitu(zt, zs, zh):
     * From the NEMO `eos_insitu` docstring
 
         >>> round(eos_insitu(40., 40., 10000.), 5)
-        1060.93299
+        np.float64(1060.93299)
 
     * From Jackett and McDougall (1995) [eos_insitu_1]_
 
         >>> round(eos_insitu(3., 35.5, 3000.), 3)
-        1041.833
+        np.float64(1041.833)
 
     Notes
     -----
